pkg/endpoint: move lock functions to lock.go
This is a code organization change; no functional changes are present.

Signed-off by: Ian Vernon <ian@cilium.io>
